+++ src/main/java/insane96mcp/survivalreimagined/module/experience/enchanting/EnsorcellerBlockEntity.java
@@ -1,5 +1,6 @@
 package insane96mcp.survivalreimagined.module.experience.enchanting;
 
+import insane96mcp.survivalreimagined.module.experience.GlobalExperience;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.NonNullList;
@@ -9,12 +10,15 @@ import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
 import net.minecraft.world.Container;
 import net.minecraft.world.ContainerHelper;
 import net.minecraft.world.WorldlyContainer;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.ExperienceOrb;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.ContainerData;
 import net.minecraft.world.inventory.ContainerLevelAccess;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BaseContainerBlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
@@ -146,6 +150,14 @@ public class EnsorcellerBlockEntity extends BaseContainerBlockEntity implements 
     @Override
     public void clearContent() {
         this.items.clear();
+    }
+
+    public void dropExperience(Level level) {
+        ExperienceOrb experienceOrb = EntityType.EXPERIENCE_ORB.create(level);
+        experienceOrb.setPos(this.getBlockPos().getX() + 0.5d, this.getBlockPos().getY() + 0.5d, this.getBlockPos().getZ() + 0.5d);
+        experienceOrb.getPersistentData().putBoolean(GlobalExperience.XP_PROCESSED, true);
+        experienceOrb.value = this.rollsPerformed * 5;
+        level.addFreshEntity(experienceOrb);
     }
 
     public @Nullable ClientboundBlockEntityDataPacket getUpdatePacket() {