+++ changelog.md
@@ -23,6 +23,7 @@
 * Slight increase to Unbreaking formula (x1.29, x1.67, x2.11, x2.5 -> x1.33, x1.82, x2.5, x3.33 durability at levels I to IV)
 * Slightly decreased Durium Ore XP drop
 * Blasting is now affected by Water, midair, haste and mining fatigue
+* Breaking an ensorceller now gives back some experience used to roll
 * Pillagers now drop emeralds and arrows
 * Phantoms now actually spawn in The End
 * Fixed datapacks not disabling if they were already enabled in a world